+++ src/main/gen/icu/windea/pls/script/psi/ParadoxScriptInlineMathUnaryExpression.java
@@ -1,9 +1,12 @@
 // This is a generated file. Not intended for manual editing.
 package icu.windea.pls.script.psi;
 
-import com.intellij.navigation.*;
-import com.intellij.psi.search.*;
+import java.util.List;
 import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+import com.intellij.navigation.ItemPresentation;
+import com.intellij.psi.search.GlobalSearchScope;
+import com.intellij.psi.search.SearchScope;
 
 public interface ParadoxScriptInlineMathUnaryExpression extends ParadoxScriptInlineMathExpression {