+++ src/main/gen/icu/windea/pls/script/psi/ParadoxScriptParameterCondition.java
@@ -1,13 +1,15 @@
 // This is a generated file. Not intended for manual editing.
 package icu.windea.pls.script.psi;
 
-import com.intellij.navigation.*;
-import com.intellij.psi.*;
-import com.intellij.psi.search.*;
+import java.util.List;
 import org.jetbrains.annotations.*;
-
-import javax.swing.*;
-import java.util.*;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiListLikeElement;
+import com.intellij.navigation.ItemPresentation;
+import com.intellij.openapi.util.Iconable.IconFlags;
+import com.intellij.psi.search.GlobalSearchScope;
+import com.intellij.psi.search.SearchScope;
+import javax.swing.Icon;
 
 public interface ParadoxScriptParameterCondition extends PsiListLikeElement {